refactor: fix typos
