+++ packages/compass-e2e-tests/helpers/read-first-document-content.ts
